docs/requirements.txt
evdspy>=1.1.38
sphinx_rtd_theme
sphinx_book_theme
